requirements .txt
fastapi
uvicorn
python-dotenv
youtube-transcript-api
langchain
langchain_community
langchain-google-genai
faiss-cpu
